Add auth ajax functions
Added auth ajax functions in api file. Moved auth events functions to
events file.
