+++ assets/scripts/auth/events.js
@@ -1,0 +1,46 @@
+const getFormFields = require('../../../lib/get-form-fields')
+const api = require('./api')
+const ui = require('./ui')
+
+const onSignUp = event => {
+  const form = event.target
+  event.preventDefault()
+  const formData = getFormFields(form)
+  api.signUp(formData)
+    .then(ui.onSignUpSuccess)
+    .catch(ui.onSignUpFailure)
+}
+
+const onSignIn = event => {
+  event.preventDefault()
+  const form = event.target
+  const formData = getFormFields(form)
+  api.signIn(formData)
+    .then(ui.onSignInSuccess)
+    .catch(ui.onSignInFailure)
+}
+
+const onSignOut = event => {
+  event.preventDefault()
+  const form = event.target
+  const formData = getFormFields(form)
+  api.signOut(formData)
+    .then(ui.onSignOutSuccess)
+    .catch(ui.onSignOutFailure)
+}
+
+const onChangePassword = event => {
+  event.preventDefault()
+  const form = event.target
+  const formData = getFormFields(form)
+  api.changePassword(formData)
+    .then(ui.onChangePasswordSuccess)
+    .catch(ui.onChangePasswordFailure)
+}
+
+module.exports = {
+  onSignUp,
+  onSignIn,
+  onSignOut,
+  onChangePassword
+}
