+++ assets/scripts/bucketlist/events.js
@@ -2,42 +2,6 @@ const getFormFields = require('../../../lib/get-form-fields')
 const api = require('./api')
 const ui = require('./ui')
 
-const onSignUp = event => {
-  const form = event.target//
-  event.preventDefault()
-  const formData = getFormFields(form)
-  api.signUp(formData)
-    .then(ui.onSignUpSuccess)
-    .catch(ui.onSignUpFailure)
-}
-
-const onSignIn = event => {
-  event.preventDefault()
-  const form = event.target
-  const formData = getFormFields(form)
-  api.signIn(formData)
-    .then(ui.onSignInSuccess)
-    .catch(ui.onSignInFailure)
-}
-
-const onSignOut = event => {
-  event.preventDefault()
-  const form = event.target
-  const formData = getFormFields(form)
-  api.signOut(formData)
-    .then(ui.onSignOutSuccess)
-    .catch(ui.onSignOutFailure)
-}
-
-const onChangePassword = event => {
-  event.preventDefault()
-  const form = event.target
-  const formData = getFormFields(form)
-  api.changePassword(formData)
-    .then(ui.onChangePasswordSuccess)
-    .catch(ui.onChangePasswordFailure)
-}
-
 const onAddListItem = event => {
   event.preventDefault()
   const form = event.target
@@ -71,13 +35,9 @@ const addHandlers = () => {
 }
 
 module.exports = {
-  onSignUp,
-  onSignIn,
-  onSignOut,
-  onChangePassword,
   onAddListItem,
   onDeleteListItem,
   onUpdateListItem,
   onGetListItems,
-  addHandlers,
+  addHandlers
 }
